# Persist addresses as diff instead of full snapshots