--- a/lib/core/src/Cardano/Wallet/Primitive/AddressDerivation/Byron.hs
+++ b/lib/core/src/Cardano/Wallet/Primitive/AddressDerivation/Byron.hs
@@ -28,6 +28,7 @@
 module Cardano.Wallet.Primitive.AddressDerivation.Byron
     ( -- * Types
       ByronKey(..)
+    , DerivationPath
 
       -- * Generation
     , unsafeGenerateKeyFromSeed