--- a/lib/core/test/bench/db/Main.hs
+++ b/lib/core/test/bench/db/Main.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE TupleSections #-}
 {-# LANGUAGE TypeApplications #-}
 
 {-# OPTIONS_GHC -fno-warn-orphans #-}
@@ -52,17 +53,23 @@ import Cardano.Startup
 import Cardano.Wallet.DB
     ( DBLayer (..), PrimaryKey (..), cleanDB )
 import Cardano.Wallet.DB.Sqlite
-    ( DefaultFieldValues (..), newDBLayer )
+    ( DefaultFieldValues (..), PersistState, newDBLayer )
 import Cardano.Wallet.DummyTarget.Primitive.Types
     ( block0, dummyGenesisParameters, dummyProtocolParameters, mkTxId )
 import Cardano.Wallet.Primitive.AddressDerivation
     ( DelegationAddress (..)
     , Depth (..)
     , NetworkDiscriminant (..)
+    , Passphrase (..)
+    , PersistPrivateKey
     , WalletKey (..)
     )
+import Cardano.Wallet.Primitive.AddressDerivation.Byron
+    ( ByronKey )
 import Cardano.Wallet.Primitive.AddressDerivation.Jormungandr
     ( JormungandrKey (..), generateKeyFromSeed, unsafeGenerateKeyFromSeed )
+import Cardano.Wallet.Primitive.AddressDiscovery.Random
+    ( DerivationPath, RndState (..), mkRndState )
 import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
     ( AddressPool
     , SeqState (..)
@@ -79,6 +86,7 @@ import Cardano.Wallet.Primitive.Types
     ( ActiveSlotCoefficient (..)
     , ActiveSlotCoefficient (..)
     , Address (..)
+    , Block (..)
     , BlockHeader (..)
     , Coin (..)
     , Direction (..)
@@ -128,6 +136,8 @@ import Data.Functor
     ( ($>) )
 import Data.Functor.Identity
     ( Identity (..) )
+import Data.Map.Strict
+    ( Map )
 import Data.Maybe
     ( fromMaybe )
 import Data.Proxy
@@ -157,6 +167,7 @@ import System.IO.Unsafe
 import System.Random
     ( mkStdGen, randoms )
 
+import qualified Cardano.Wallet.Primitive.AddressDerivation.Byron as Byron
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Char8 as B8
 import qualified Data.Map.Strict as Map
@@ -166,7 +177,8 @@ main = withUtf8Encoding $ do
     defaultMain
         [ withDB bgroupWriteUTxO
         , withDB bgroupReadUTxO
-        , withDB bgroupSeqState
+        , withDB bgroupWriteSeqState
+        , withDB bgroupWriteRndState
         , withDB bgroupWriteTxHistory
         , withDB bgroupReadTxHistory
         ]
@@ -189,19 +201,18 @@ bgroupWriteUTxO db = bgroup "UTxO (Write)"
     --
     --      #Checkpoints   UTxO Size
     [ bUTxO            1           0
-    , bUTxO           10          10
-    , bUTxO           10         100
+    , bUTxO            1          10
+    , bUTxO            1         100
+    , bUTxO            1        1000
+    , bUTxO            1       10000
+    , bUTxO            1      100000
     , bUTxO           10        1000
-    , bUTxO           10       10000
-    , bUTxO          100           0
-    , bUTxO          100          10
-    , bUTxO          100         100
     , bUTxO          100        1000
-    , bUTxO          100       10000
     ]
   where
-    bUTxO n s = bench lbl $ withCleanDB db $ benchPutUTxO n s
-        where lbl = n|+" CP x "+|s|+" UTxO"
+    bUTxO n s = bench lbl $ withCleanDB db walletFixture $
+        benchPutUTxO n s . fst
+      where lbl = n|+" CP x "+|s|+" UTxO"
 
 bgroupReadUTxO :: DBLayerBench -> Benchmark
 bgroupReadUTxO db = bgroup "UTxO (Read)"
@@ -217,24 +228,139 @@ bgroupReadUTxO db = bgroup "UTxO (Read)"
     bUTxO n s = bench lbl $ withUTxO db n s benchReadUTxO
         where lbl = n|+" CP x "+|s|+" UTxO"
 
+benchPutUTxO :: Int -> Int -> DBLayerBench -> IO ()
+benchPutUTxO numCheckpoints utxoSize DBLayer{..} = do
+    let cps = mkCheckpoints numCheckpoints utxoSize
+    unsafeRunExceptT $ mapExceptT atomically $ mapM_ (putCheckpoint testPk) cps
+
+mkCheckpoints :: Int -> Int -> [WalletBench]
+mkCheckpoints numCheckpoints utxoSize =
+    [ force (cp i) | !i <- [1..numCheckpoints] ]
+  where
+    cp i = unsafeInitWallet
+        (UTxO (utxo i))
+        (BlockHeader
+            (SlotNo $ fromIntegral i)
+            (Quantity $ fromIntegral i)
+            (Hash $ label "parentHeaderHash" i)
+            (Hash $ label "headerHash" i)
+        )
+        initDummySeqState
+        dummyGenesisParameters
+
+    utxo i = Map.fromList $ zip
+        (map fst $ mkInputs i utxoSize)
+        (mkOutputs i utxoSize)
+
+benchReadUTxO :: DBLayerBench -> IO (Maybe WalletBench)
+benchReadUTxO DBLayer{..} = atomically $ readCheckpoint testPk
+
+-- Set up a database with some UTxO in checkpoints.
+withUTxO
+    :: NFData b
+    => DBLayerBench
+    -> Int
+    -> Int
+    -> (DBLayerBench -> IO b)
+    -> Benchmarkable
+withUTxO db n s = perRunEnv (utxoFixture db n s $> db)
+
+utxoFixture :: DBLayerBench -> Int -> Int -> IO ()
+utxoFixture db@DBLayer{..} numCheckpoints utxoSize = do
+    walletFixture db
+    let cps = mkCheckpoints numCheckpoints utxoSize
+    unsafeRunExceptT $ mapM_ (mapExceptT atomically . putCheckpoint testPk) cps
+
 ----------------------------------------------------------------------------
 -- Wallet State (Sequential Scheme) Benchmarks
 --
-bgroupSeqState :: DBLayerBench -> Benchmark
-bgroupSeqState db = bgroup "SeqState"
+bgroupWriteSeqState :: DBLayerBench -> Benchmark
+bgroupWriteSeqState db = bgroup "SeqState"
     --      #Checkpoints  #Addresses
-    [ bSeqState       10          10
-    , bSeqState       10         100
+    [ bSeqState        1          10
+    , bSeqState        1         100
+    , bSeqState        1        1000
+    , bSeqState        1       10000
+    , bSeqState        1      100000
     , bSeqState       10        1000
-    , bSeqState       10       10000
-    , bSeqState      100          10
-    , bSeqState      100         100
     , bSeqState      100        1000
-    , bSeqState      100       10000
     ]
   where
-    bSeqState n a = bench lbl $ withCleanDB db $ benchPutSeqState n a
-        where lbl = n|+" CP x "+|a|+" addr"
+    bSeqState n a = bench lbl $ withCleanDB db fixture (uncurry benchPutSeqState)
+      where
+        lbl = n|+" CP x "+|a|+" addr"
+        fixture db_ = do
+            walletFixture db_
+            pure cps
+        cps :: [WalletBench]
+        cps =
+            [ snd $ initWallet (withMovingSlot i block0) dummyGenesisParameters $
+                SeqState
+                    (mkPool a i)
+                    (mkPool a i)
+                    emptyPendingIxs
+                    rewardAccount
+            | i <- [1..n]
+            ]
+
+benchPutSeqState :: DBLayerBench -> [WalletBench] -> IO ()
+benchPutSeqState DBLayer{..} cps = do
+    unsafeRunExceptT $ mapExceptT atomically $ mapM_ (putCheckpoint testPk) cps
+
+mkPool
+    :: forall c. (Typeable c)
+    => Int -> Int -> AddressPool c JormungandrKey
+mkPool numAddrs i = mkAddressPool ourAccount defaultAddressPoolGap addrs
+  where
+    addrs = [ force (mkAddress i j) | j <- [1..numAddrs] ]
+
+----------------------------------------------------------------------------
+-- Wallet State (Random Scheme) Benchmarks
+--
+bgroupWriteRndState :: DBLayerBenchByron -> Benchmark
+bgroupWriteRndState db = bgroup "RndState"
+    --      #Checkpoints  #Addresses
+    [ bRndState        1          10
+    , bRndState        1         100
+    , bRndState        1        1000
+    , bRndState        1       10000
+    , bRndState        1      100000
+    , bRndState       10        1000
+    , bRndState      100        1000
+    ]
+  where
+    bRndState n a = bench lbl $ withCleanDB db fixture (uncurry benchPutRndState)
+      where
+        lbl = n|+" CP x "+|a|+" addr"
+        fixture db_ = do
+            walletFixtureByron db_
+            pure cps
+        cps :: [Wallet (RndState 'Mainnet)]
+        cps =
+            [ snd $ initWallet (withMovingSlot i block0) dummyGenesisParameters $
+                RndState
+                    { hdPassphrase = dummyPassphrase
+                    , accountIndex = minBound
+                    , addresses = mkRndAddresses a i
+                    , pendingAddresses = mempty
+                    , gen = mkStdGen 42
+                    }
+            | i <- [1..n]
+            ]
+
+benchPutRndState
+    :: DBLayerBenchByron
+    -> [Wallet (RndState 'Mainnet)]
+    -> IO ()
+benchPutRndState DBLayer{..} cps =
+    unsafeRunExceptT $ mapExceptT atomically $ mapM_ (putCheckpoint testPk) cps
+
+mkRndAddresses
+    :: Int -> Int -> Map DerivationPath Address
+mkRndAddresses numAddrs i =
+    Map.fromList addrs
+  where
+    addrs = [ force ((toEnum i, toEnum j), mkAddress i j) | j <- [1..numAddrs] ]
 
 ----------------------------------------------------------------------------
 -- Tx history Benchmarks
@@ -301,7 +427,8 @@ bgroupWriteTxHistory db = bgroup "TxHistory (Write)"
     ]
   where
     bTxHistory n i o r =
-        bench lbl $ withCleanDB db $ benchPutTxHistory n i o r
+        bench lbl $ withCleanDB db walletFixture $
+            benchPutTxHistory n i o r . fst
       where
         lbl = n|+" w/ "+|i|+"i + "+|o|+"o ["+|inf|+".."+|sup|+"]"
         inf = head r
@@ -334,70 +461,6 @@ bgroupReadTxHistory db = bgroup "TxHistory (Read)"
             (Just inf, Nothing) -> inf|+".."
             (Nothing, Just sup) -> ".."+|sup|+""
             (Just inf, Just sup) -> inf|+".."+|sup|+""
-
-----------------------------------------------------------------------------
--- Criterion env functions for database setup
-
--- | Sets up a benchmark environment with the SQLite DBLayer using a file
--- database in a temporary location.
-withDB :: (DBLayerBench -> Benchmark) -> Benchmark
-withDB bm = envWithCleanup setupDB cleanupDB (\ ~(_, _, db) -> bm db)
-
-setupDB :: IO (FilePath, SqliteContext, DBLayerBench)
-setupDB = do
-    f <- emptySystemTempFile "bench.db"
-    (ctx, db) <- newDBLayer nullTracer defaultFieldValues (Just f) ti
-    pure (f, ctx, db)
-  where
-    ti = pure . runIdentity . singleEraInterpreter (GenesisParameters
-        { getGenesisBlockHash = Hash $ BS.replicate 32 0
-        , getGenesisBlockDate = StartTime $ posixSecondsToUTCTime 0
-        , getSlotLength = SlotLength 1
-        , getEpochLength = EpochLength 21600
-        , getEpochStability = Quantity 108
-        , getActiveSlotCoefficient = ActiveSlotCoefficient 1
-        })
-
-defaultFieldValues :: DefaultFieldValues
-defaultFieldValues = DefaultFieldValues
-    { defaultActiveSlotCoefficient = ActiveSlotCoefficient 1.0
-    , defaultDesiredNumberOfPool = 50
-    , defaultMinimumUTxOValue = Coin 0
-    , defaultHardforkEpoch = Nothing
-        -- NOTE value in the genesis when at the time this migration was needed.
-    }
-
-cleanupDB :: (FilePath, SqliteContext, DBLayerBench) -> IO ()
-cleanupDB (db, ctx, _) = do
-    handle (\SqliteException{} -> pure ()) $ destroyDBLayer ctx
-    mapM_ remove [db, db <> "-shm", db <> "-wal"]
-  where
-    remove f = doesFileExist f >>= \case
-        True -> removeFile f
-        False -> pure ()
-
--- | Cleans the database before running the benchmark.
--- It also cleans the database after running the benchmark. That is just to
--- exercise the delete functions.
-withCleanDB
-    :: NFData b
-    => DBLayerBench
-    -> (DBLayerBench -> IO b)
-    -> Benchmarkable
-withCleanDB db = perRunEnv (walletFixture db $> db)
-
-walletFixture :: DBLayerBench -> IO ()
-walletFixture db@DBLayer{..} = do
-    cleanDB db
-    atomically $ unsafeRunExceptT $ initializeWallet
-        testPk
-        testCp
-        testMetadata
-        mempty
-        dummyProtocolParameters
-
-----------------------------------------------------------------------------
--- TxHistory benchmarks
 
 benchPutTxHistory
     :: Int
@@ -480,72 +543,88 @@ txHistoryFixture db@DBLayer{..} bSize range = do
     atomically $ unsafeRunExceptT $ putTxHistory testPk txs
 
 ----------------------------------------------------------------------------
--- UTxO benchmarks
+-- Criterion env functions for database setup
 
-benchPutUTxO :: Int -> Int -> DBLayerBench -> IO ()
-benchPutUTxO numCheckpoints utxoSize DBLayer{..} = do
-    let cps = mkCheckpoints numCheckpoints utxoSize
-    unsafeRunExceptT $ mapExceptT atomically $ mapM_ (putCheckpoint testPk) cps
-
-mkCheckpoints :: Int -> Int -> [WalletBench]
-mkCheckpoints numCheckpoints utxoSize =
-    [ force (cp i) | !i <- [1..numCheckpoints] ]
-  where
-    cp i = unsafeInitWallet
-        (UTxO (utxo i))
-        (BlockHeader
-            (SlotNo $ fromIntegral i)
-            (Quantity $ fromIntegral i)
-            (Hash $ label "parentHeaderHash" i)
-            (Hash $ label "headerHash" i)
+-- | Sets up a benchmark environment with the SQLite DBLayer using a file
+-- database in a temporary location.
+withDB
+    :: forall s k.
+        ( PersistState s
+        , PersistPrivateKey (k 'RootK)
         )
-        initDummyState
-        dummyGenesisParameters
+    => (DBLayer IO s k -> Benchmark)
+    -> Benchmark
+withDB bm = envWithCleanup setupDB cleanupDB (\ ~(_, _, db) -> bm db)
 
-    utxo i = Map.fromList $ zip
-        (map fst $ mkInputs i utxoSize)
-        (mkOutputs i utxoSize)
-
-benchReadUTxO :: DBLayerBench -> IO (Maybe WalletBench)
-benchReadUTxO DBLayer{..} = atomically $ readCheckpoint testPk
-
--- Set up a database with some UTxO in checkpoints.
-withUTxO
-    :: NFData b
-    => DBLayerBench
-    -> Int
-    -> Int
-    -> (DBLayerBench -> IO b)
-    -> Benchmarkable
-withUTxO db n s = perRunEnv (utxoFixture db n s $> db)
-
-utxoFixture :: DBLayerBench -> Int -> Int -> IO ()
-utxoFixture db@DBLayer{..} numCheckpoints utxoSize = do
-    walletFixture db
-    let cps = mkCheckpoints numCheckpoints utxoSize
-    unsafeRunExceptT $ mapM_ (mapExceptT atomically . putCheckpoint testPk) cps
-
-----------------------------------------------------------------------------
--- SeqState Address Discovery
-
-benchPutSeqState :: Int -> Int -> DBLayerBench -> IO ()
-benchPutSeqState numCheckpoints numAddrs DBLayer{..} =
-    unsafeRunExceptT $ mapExceptT atomically $ mapM_ (putCheckpoint testPk)
-        [ snd $ initWallet block0 dummyGenesisParameters $
-            SeqState
-                (mkPool numAddrs i)
-                (mkPool numAddrs i)
-                emptyPendingIxs
-                rewardAccount
-        | i <- [1..numCheckpoints]
-        ]
-
-mkPool
-    :: forall c. (Typeable c)
-    => Int -> Int -> AddressPool c JormungandrKey
-mkPool numAddrs i = mkAddressPool ourAccount defaultAddressPoolGap addrs
+setupDB
+    :: forall s k.
+        ( PersistState s
+        , PersistPrivateKey (k 'RootK)
+        )
+    => IO (FilePath, SqliteContext, DBLayer IO s k)
+setupDB = do
+    f <- emptySystemTempFile "bench.db"
+    (ctx, db) <- newDBLayer nullTracer defaultFieldValues (Just f) ti
+    pure (f, ctx, db)
   where
-    addrs = [ force (mkAddress i j) | j <- [1..numAddrs] ]
+    ti = pure . runIdentity . singleEraInterpreter (GenesisParameters
+        { getGenesisBlockHash = Hash $ BS.replicate 32 0
+        , getGenesisBlockDate = StartTime $ posixSecondsToUTCTime 0
+        , getSlotLength = SlotLength 1
+        , getEpochLength = EpochLength 21600
+        , getEpochStability = Quantity 108
+        , getActiveSlotCoefficient = ActiveSlotCoefficient 1
+        })
+
+defaultFieldValues :: DefaultFieldValues
+defaultFieldValues = DefaultFieldValues
+    { defaultActiveSlotCoefficient = ActiveSlotCoefficient 1.0
+    , defaultDesiredNumberOfPool = 50
+    , defaultMinimumUTxOValue = Coin 0
+    , defaultHardforkEpoch = Nothing
+        -- NOTE value in the genesis when at the time this migration was needed.
+    }
+
+cleanupDB :: (FilePath, SqliteContext, DBLayer IO s k) -> IO ()
+cleanupDB (db, ctx, _) = do
+    handle (\SqliteException{} -> pure ()) $ destroyDBLayer ctx
+    mapM_ remove [db, db <> "-shm", db <> "-wal"]
+  where
+    remove f = doesFileExist f >>= \case
+        True -> removeFile f
+        False -> pure ()
+
+-- | Cleans the database before running the benchmark.
+-- It also cleans the database after running the benchmark. That is just to
+-- exercise the delete functions.
+withCleanDB
+    :: NFData fixture
+    => DBLayer IO s k
+    -> (DBLayer IO s k -> IO fixture)
+    -> ((DBLayer IO s k, fixture) -> IO ())
+    -> Benchmarkable
+withCleanDB db fixture =
+    perRunEnv $ (db,) <$> fixture db
+
+walletFixture :: DBLayerBench -> IO ()
+walletFixture db@DBLayer{..} = do
+    cleanDB db
+    atomically $ unsafeRunExceptT $ initializeWallet
+        testPk
+        testCp
+        testMetadata
+        mempty
+        dummyProtocolParameters
+
+walletFixtureByron :: DBLayerBenchByron -> IO ()
+walletFixtureByron db@DBLayer{..} = do
+    cleanDB db
+    atomically $ unsafeRunExceptT $ initializeWallet
+        testPk
+        testCpByron
+        testMetadata
+        mempty
+        dummyProtocolParameters
 
 ----------------------------------------------------------------------------
 -- Disk space usage tests
@@ -626,7 +705,9 @@ benchDiskSize action = bracket setupDB cleanupDB $ \(f, ctx, db) -> do
 -- Mock data to use for benchmarks
 
 type DBLayerBench = DBLayer IO (SeqState 'Mainnet JormungandrKey) JormungandrKey
+type DBLayerBenchByron = DBLayer IO (RndState 'Mainnet) ByronKey
 type WalletBench = Wallet (SeqState 'Mainnet JormungandrKey)
+type WalletBenchByron = Wallet (RndState 'Mainnet)
 
 instance NFData (DBLayer m s k) where
     rnf _ = ()
@@ -635,17 +716,29 @@ instance NFData SqliteContext where
     rnf _ = ()
 
 testCp :: WalletBench
-testCp = snd $ initWallet block0 dummyGenesisParameters initDummyState
+testCp = snd $ initWallet block0 dummyGenesisParameters initDummySeqState
 
-{-# NOINLINE initDummyState #-}
-initDummyState :: SeqState 'Mainnet JormungandrKey
-initDummyState =
+testCpByron :: WalletBenchByron
+testCpByron = snd $ initWallet block0 dummyGenesisParameters initDummyRndState
+
+{-# NOINLINE initDummySeqState #-}
+initDummySeqState :: SeqState 'Mainnet JormungandrKey
+initDummySeqState =
     mkSeqStateFromRootXPrv (xprv, mempty) defaultAddressPoolGap
   where
     mnemonic = unsafePerformIO
         $ SomeMnemonic . entropyToMnemonic @15
         <$> genEntropy @(EntropySize 15)
     xprv = generateKeyFromSeed (mnemonic, Nothing) mempty
+
+{-# NOINLINE initDummyRndState #-}
+initDummyRndState :: RndState 'Mainnet
+initDummyRndState =
+    mkRndState rootK 42
+  where
+    rootK = Byron.generateKeyFromSeed mnemonic mempty
+    mnemonic = unsafePerformIO $
+        SomeMnemonic . entropyToMnemonic @12 <$> genEntropy @(EntropySize 12)
 
 testMetadata :: WalletMetadata
 testMetadata = WalletMetadata
@@ -674,6 +767,18 @@ rewardAccount = publicKey $ unsafeGenerateKeyFromSeed (seed, Nothing) mempty
 -- | Make a prefixed bytestring for use as a Hash or Address.
 label :: Show n => String -> n -> B8.ByteString
 label prefix n = B8.take 32 $ B8.pack (prefix <> show n) <> B8.replicate 32 '0'
+
+dummyPassphrase :: Passphrase any
+dummyPassphrase = Passphrase "dummy-passphrase"
+
+-- | Make sure to generate
+withMovingSlot :: Int -> Block -> Block
+withMovingSlot i b@(Block h _ _) = b
+    { header = h
+        { slotNo = SlotNo (fromIntegral i)
+        , blockHeight = Quantity (fromIntegral i)
+        }
+    }
 
 mkAddress :: Int -> Int -> Address
 mkAddress i j =