--- a/stack.yaml
+++ b/stack.yaml
@@ -27,6 +27,9 @@ extra-deps:
 # 'zip' with an extra flag to disable bzlib2 library
 - zip-1.3.0
 
+# Needed for recently introduced support of TLS-1.3
+- tls-1.5.4
+
 # persistent-2.10.2 with CASCADE DELETE support for SQLite.
 #
 # See: https://github.com/input-output-hk/persistent/tree/cardano-wallet