… lib/core/test/unit/Cardano/Wallet/DB/Arbitrary.hs
… lib/core/test/unit/Cardano/Wallet/DB/StateMachine.hs
… lib/core/test/unit/Cardano/Wallet/Primitive/AddressDiscovery/SequentialSpec.hs